requirements.txt
# X API
tweepy>=4.14.0

# LLM APIs
openai>=1.12.0
anthropic>=0.18.0

# RAG / Vector Store
pinecone>=3.0.0

# Embeddings
sentence-transformers>=2.3.0

# Scheduling
APScheduler>=3.10.0
pytz>=2023.3

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Utilities
httpx>=0.26.0
tenacity>=8.2.0

# Dashboard API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
websockets>=12.0
